Throw correct exceptions from collapse/selectAllChildren on detached doctypes
According to the previous spec, trying to collapse to or select a
doctype that is not in the current document would not throw.  But
Chrome, Firefox, and Edge all do throw.  (WebKit doesn't seem to throw
for doctypes at all.)  The wpt test also tests that browsers do throw in
this case.