Throw exception for removeRange on range that's not in selection
This matches Firefox, which was historically the only engine to ever
have a working implementation of removeRange.  WebKit still has none,
Edge seems to be a no-op, and Chrome just added theirs in February.  If
there was only ever one implementation of the feature, the spec should
follow that implementation.

(Firefox doesn't actually throw a DOMException here, it throws a
nonstandard exception.  That's a side point, though.)